Add my Prototype from Unity
In this Unit, I programed an arcade-style Sumo battle with the objective of knocking increasingly difficult waves of enemies off of a floating island, using power ups to help defeat them. In creating this prototype, I learned how to implement new gameplay mechanics into  projects, which are new rules or systems that make the game more interesting to play. On one hand, I learned to program a powerup, which give the player a temporary advantage. On the other hand, I learned how to program increasingly difficult enemy waves, which make survival more challenging for the player. A good balance of powerups and increasing difficulty make for a much more interesting gameplay experience.